Fix stop button
I'm not exactly sure what about the SIGTERM did not work.
Even assuming it killed the right thing,
a lot of instructions could still be buffered,
possibly even the whole file.

I chose to transfer 1 byte at a time in NodeJS directly,
so we have full control over the flow.

This is not tested on hardware yet.
Possibly this could make the whole thing slower?
In that case we could play with the buffer sizes,
at the expense of stopping later.