Logic improvements
Stop button is only enabled if a print is in progress.

Plotter is set as busy asap to prevent race conditions.

Close the reader on error instead of the writer,
as the former has all the recovering logic.

Status is refreshed after each HTTP request where it makes sense to do
so. Helps with fluidity.

If you're not the author of the last/current document,
the status will only be shown as busy if running.
Not sure if it changes much from the previous logic,
but I find it more readable.