Add pause/resume button
I wish .pause() would work as I would expect it to work,
but no, it's more complicated than that.
So complicated that I'm not sure how it works actually.
But it does, and even if the necessary .pause()
is just after .unpipe() which looks scary,
it doesn't seem to miss a byte from my testing.